refactor : repository 인터페이스와 구현체로 관리할 수 있도록 수정
